Backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
# Frontend dependencies
react==17.0.2
react-dom==17.0.2
react-scripts==5.0.0
react-router-dom==6.2.1
axios==0.24.0
bootstrap==5.3.0
react-bootstrap==2.0.0
react-icons==4.3.1
react-toastify==8.0.3
react-quill==1.3.5
react-markdown==7.0.1
react-syntax-highlighter==15.4.3
react-syntax-highlighter-lazy-load==1.0.0
react-syntax-highlighter-line-number==1.0.0
react-syntax-highlighter-line-number-lazy-load==1.0.0
react-syntax-highlighter-prism==1.0.0
react-syntax-highlighter-prism-lazy-load==1.0.0
react-syntax-highlighter-prism-line-number==1.0.0
react-hook-form==7.0.0
react-hot-toast==1.0.4
